refactor: update error handling and data structure types for consistency

+++ include/data_structures/stack.h
@@ -2,6 +2,9 @@
 
 #include <stddef.h>
 
+#include <stdint.h>
+
+
 #define STACK_EMPTY -1
 #define STACK_OK 0
 #define STACK_ERR_ALLOC 1
@@ -10,14 +13,14 @@
 
 typedef struct StackDef {
     void* data;
-    size_t top;
-    size_t total_elements;
-    size_t member_size;
+    int64_t top;
+    int64_t total_elements;
+    int64_t member_size;
 } Stack;
 
-Stack* create_stack(size_t member_size, size_t total_elements);
+Stack* create_stack(int64_t member_size, int64_t total_elements);
 bool push_to_stack(Stack* stack, void* data);
 bool expand_stack(Stack* stack);
 bool pop_from_stack(Stack* stack, void* target);
 void reverse_stack(Stack* stack);
-bool get_stack_index(Stack* stack, size_t index, void* target);
+bool get_stack_index(Stack* stack, int64_t index, void* target);
